+++ src/components/InputText/types.d.ts
@@ -1,0 +1,7 @@
+export type InputTextProps = {
+  label: string;
+  warningLabel: string;
+  placeholder?: string;
+  error?: boolean;
+  onChange?: (event: React.FormEvent<HTMLInputElement>) => void;
+};