+++ CHANGELOG.md
@@ -2,7 +2,15 @@
 
 This is just an example changelog used to cover some of the functionalities that are been added to this challenge project.
 
-### To Publish
+### Version 2
+
+- **Feature 7:** Add more Tag and also Breweries in User Store.
+
+- **Feature 6:** Lazy Loading in Breweries Page
+
+- **Fix 5:** Visual Feedback error when going to Breweries page and then going back to Login
+
+- **Feature 4:** Visual Feedback when User's Name is not valid in the Login Page
 
 ### Version 1